lazy fix for redis client.
