+++ lib/transport/redis/redis_client.ex
@@ -4,6 +4,9 @@ defmodule Farmbot.Transport.Redis do
   """
   use GenStage
   require Logger
+  @config Application.get_all_env(:farmbot)[:redis]
+  @ping_time 5_000
+  @save_time 90_0000
 
   @doc """
     Starts a stage for a Redis
@@ -12,14 +15,29 @@ defmodule Farmbot.Transport.Redis do
   def start_link, do: GenStage.start_link(__MODULE__, [], name: __MODULE__)
 
   def init(_) do
-    {:ok, redis} = Redis.Client.start_link
-    Process.link(redis)
-    {:consumer, redis, subscribe_to: [Farmbot.Transport]}
+    {:ok, conn} = Redix.start_link(host: "localhost", port: @config[:port])
+    Process.link(conn)
+    Process.send_after(self(), :ping, @ping_time)
+    {:consumer, conn, subscribe_to: [Farmbot.Transport]}
   end
 
   def handle_info({_from, {:status, stuff}}, redis) do
-    Redis.Client.input_value(redis, "BOT_STATUS", stuff)
+    Redis.Client.Public.input_value(redis, "BOT_STATUS", stuff)
     {:noreply, [], redis}
+  end
+
+  def handle_info(:ping, conn) do
+    Redis.Client.Public.send_redis(conn, ["PING"])
+    Process.send_after(self(), :ping, @ping_time)
+    {:noreply, [], conn}
+  end
+
+  def handle_info(:save, conn) do
+    Farmbot.System.FS.transaction fn() ->
+      Redis.Client.Public.send_redis(conn, ["SAVE"])
+    end
+    Process.send_after(self(), :save, @save_time)
+    {:noreply, [], conn}
   end
 
   def handle_info(_event, redis) do
